requirements.txt
dialogflow==0.7.2
python-dotenv==0.10.5
python-telegram-bot==11.1.0
vk-api==11.7.0
